- back-deploy 브랜치 최신 변경사항 반영

+++ back/batch/src/main/kotlin/com/metamong/infra/persistence/apartment/repository/ApartmentComplexRepositoryCustom.kt
@@ -17,15 +17,15 @@ interface ApartmentComplexRepositoryCustom {
     fun findBySidoSigunguAndNameNormalizedAndBuiltYearBetween(
         sidoSigunguCode: Int,
         nameNormalized: String,
-        builtYearFrom: Short,
-        builtYearTo: Short,
+        builtYearFrom: Int,
+        builtYearTo: Int,
     ): ApartmentComplexEntity?
 
     fun findByJibun(
         sidoSigunguCode: Int,
         eupmyeondongRiCode: Int,
         platType: PlatType,
-        bonNo: Short,
-        buNo: Short,
+        bonNo: Int,
+        buNo: Int,
     ): ApartmentComplexEntity?
 }
